[IMP] sale: Allow to invoice multiple sale returns
When two or more orders have returns after been invoice and a user
wants to make a single credit note for them all, the default values in
the sale lot invoincing wizard makes it impossible as they're hidden for
the user. So we default to 'all' if more than one orders are selected.
Otherwise the behavior remains the same.